Updates docs template generator
