+++ docs/index.880915d0.js
@@ -1,0 +1,7 @@
+function t(t,e,n,a){Object.defineProperty(t,e,{get:n,set:a,enumerable:!0,configurable:!0})}var e=globalThis,n={},a={},o=e.parcelRequire0b36;null==o&&((o=function(t){if(t in n)return n[t].exports;if(t in a){var e=a[t];delete a[t];var o={id:t,exports:{}};return n[t]=o,e.call(o.exports,o,o.exports),o.exports}var r=new Error("Cannot find module '"+t+"'");throw r.code="MODULE_NOT_FOUND",r}).register=function(t,e){a[t]=e},e.parcelRequire0b36=o),(0,o.register)("aES5g",(function(e,n){function a(){Element.prototype.setDataAttr=function(t,e=""){return this.setAttribute(`data-${t}`,e)}}function o(){Element.prototype.removeDataAttr=function(t){return this.removeAttribute(`data-${t}`)}}function r(){Element.prototype.getDataAttr=function(t,e){return e?this.getAttribute(`data-${t}="${e}"`):this.getAttribute(`data-${t}`)}}function i(){Element.prototype.hasDataAttr=function(t,e){return e?this.hasAttribute(`data-${t}="${e}"`):this.hasAttribute(`data-${t}`)}}function s(){Element.prototype.closestDataAttr=function(t,e){return e?this.closest(`[data-${t}="${e}"]`):this.closest(`[data-${t}]`)}}function l(){Element.prototype.matchesDataAttr=function(t,e){return e?this.matches(`[data-${t}="${e}"]`):this.matches(`[data-${t}]`)}}function c(){Element.prototype.toggleDataAttr=function(t,e){return this.hasDataAttr(t)?this.removeDataAttr(t):this.setDataAttr(t,e)}}t(e.exports,"setDataAttr",(()=>a)),t(e.exports,"removeDataAttr",(()=>o)),t(e.exports,"getDataAttr",(()=>r)),t(e.exports,"hasDataAttr",(()=>i)),t(e.exports,"closestDataAttr",(()=>s)),t(e.exports,"matchesDataAttr",(()=>l)),t(e.exports,"toggleDataAttr",(()=>c))}));var r=o("aES5g");function i(...t){const e=new Set,n=new Set(["pop","tooltip",...t]);function a(){e.forEach((t=>{n.has(t.dataset.toggle)&&t.removeDataAttr("state")}))}function o(t){const e=t.closestDataAttr("dropbox"),n=t.hasDataAttr("state");e||n||a(),t.toggleDataAttr("state","active")}return document.addEventListener("click",(t=>{const n=t.target.closestDataAttr("toggle");if(n){if(e.has(n))return;e.add(n),n.addEventListener("click",(t=>{o(n),t.stopPropagation()}),!0),o(n)}else a();t.stopPropagation()})),this}class s{constructor(){s.#t(),this.toggle=i}static#t(){(0,r.setDataAttr)(),(0,r.removeDataAttr)(),(0,r.getDataAttr)(),(0,r.hasDataAttr)(),(0,r.closestDataAttr)(),(0,r.matchesDataAttr)(),(0,r.toggleDataAttr)()}toggle}const l=new s;
+/* @license
+ * Jacob Perez <https://jacobxperez.github.io/blog/>
+ * Copyright (C) 2024 Jacob Perez <jacobxperez@gmx.com>
+ * Licensed under the Apache License, Version 2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
+-----------------------------------------------------------------------------*/s.prototype.vannelli={_appendString(t,e){document.querySelector(e).insertAdjacentHTML("beforeend",t)},_appendTemplate(t,e,n){const a=t.querySelector(e).content.cloneNode(!0);document.querySelector(n).appendChild(a)},_parseTemplate(t,e,n){const a=(new DOMParser).parseFromString(t,"text/html");this._appendTemplate(a,e,n)},newTemplate(t,e){const n=document.createElement("template");return n.innerHTML=t.trim(),n.content.firstElementChild,n.setAttribute("id",e),document.body.appendChild(n),this},setTemplate(t,e,n=null){return new Promise(((e,n)=>{t?e():n()})).then((()=>this._appendTemplate(document,t,e))).then((()=>{"function"==typeof n&&n()})).catch((t=>console.error(t,"Error: Template not found"))),this},fromString(t,e,n=null){return new Promise(((e,n)=>{"string"==typeof t?e():n(err="Error: Source is not a String")})).then((()=>this._appendString(t,e))).then((()=>{"function"==typeof n&&n()})).catch((t=>console.error(t))),this},fetchTemplate(t,e,n,a=null){return(async()=>{try{let o=await fetch(n),r=await o.text();this._parseTemplate(r,t,e),"function"==typeof a&&a()}catch(t){console.error(t,"Error: Template not found")}})(),this},removeTemplates(){document.querySelectorAll("template").forEach((t=>{t.remove()}))}},""===meta.title?meta.title="<h1>Rams</h1>":meta.title=`<h1>${meta.title}</h1>`;let c,d=`\n        <div id="header" data-container>\n            ${meta.title}\n        </div>\n        `,m='\n        <div data-container data-grid="main">\n            <aside id="aside" data-column="large-3 medium-3 small-4"></aside>\n            <article id="content" data-column="large-9 medium-9 small-4"></article>\n        </div>\n        ';"fullPage"===meta.type&&(m='\n        <div id="content" data-container data-grid="main"></div>\n        '),c="localhost"===location.hostname||"127.0.0.1"===location.hostname?window.location.origin+"/templates/index.59811f9a.html":window.location.origin+"/rams/templates/index.19081ad0.html",l.vannelli.newTemplate(`\n            <nav data-navbar="top">\n            </nav>\n            <header data-section="header">\n                ${d}\n            </header>\n            <main data-section="main">\n            ${m}\n            </main>\n            <footer data-section="footer">\n            </footer>\n            `,"layoutTemplate").setTemplate("#layoutTemplate","body").fetchTemplate("#navTemplate","nav",c).setTemplate("#headerTemplate","#header").setTemplate("#contentTemplate","#content",(()=>{const t=document.getElementById("aside"),e=document.getElementById("content");if("post"===meta.type){const n=e.querySelectorAll("h1, h2, h3, h4, h5, h6"),a='\n                <ul id=\'contents\' data-display="small-none">\n                    <li>\n                        <p><strong>Contents</strong></p>\n                    </li>\n                </ul>\n    \n                <ul data-display="small" data-box="border">\n                    <li>\n                        <a data-anchor data-flex data-toggle>\n                            <strong data-item="grow">Contents</strong>\n                            <span data-icon="&#xe043;"></span>\n                        </a>\n    \n                        <ul id=\'contentsDropdown\' data-dropbox>\n                        </ul>\n                    </li>\n                </ul>';t.insertAdjacentHTML("beforeend",a),n.forEach(((t,e)=>{t.setAttribute("id",`${e}`);const n=t.innerText,a=`<li><a href="#${e}">${n}</a></li>`;document.getElementById("contents").insertAdjacentHTML("beforeend",a);const o=`<li><a href="#${e}" data-anchor="menu">${n}</a></li>`;document.getElementById("contentsDropdown").insertAdjacentHTML("beforeend",o)}))}})).fetchTemplate("#footerTemplate","body > footer",c),l.toggle();
