+++ src/docs/components/index.html
@@ -10,8 +10,8 @@
         <link rel="icon" href="../favicon.svg" />
         <link rel="preload" href="../templates/index.html" as="fetch" type="text/html" crossorigin />
         <link rel="stylesheet" href="../css/index.css" />
-        <script src="../js/index.js"></script>
-        <script type="module" src="../js/defer.js"></script>
+        <script src="../js/meta.js"></script>
+        <script type="module" src="../js/index.js"></script>
         <script>
             meta.title = "Components";
             meta.type = "fullPage";
