+++ src/docs/index.html
@@ -1,4 +1,4 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="en" id="top">
     <head>
         <meta charset="UTF-8" />
@@ -14,8 +14,8 @@
         <link rel="icon" href="./favicon.svg" />
         <link rel="preload" href="templates/index.html" as="fetch" type="text/html" crossorigin />
         <link rel="stylesheet" href="css/index.css" />
-        <script src="./js/index.js"></script>
-        <script type="module" src="./js/defer.js"></script>
+        <script src="./js/meta.js"></script>
+        <script type="module" src="./js/index.js"></script>
         <script>
             meta.type = "post";
             meta.title = "Getting Started";
