+++ src/docs/js/meta.js
@@ -1,0 +1,11 @@
+/* @license
+ * Jacob Perez <https://jacobxperez.github.io/blog/>
+ * Copyright (C) 2024 Jacob Perez <jacobxperez@gmx.com>
+ * Licensed under the Apache License, Version 2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
+-----------------------------------------------------------------------------*/
+const meta = {
+    type: '',
+    title: document.title,
+    subtitle: '',
+};
