+++ src/docs/js/modules/meta.js
@@ -1,0 +1,7 @@
+const meta = {
+    type: '',
+    title: document.title,
+    subtitle: '',
+};
+
+export {meta};
